# film-contracts/openapi.json
{
  "components": {
    "schemas": {
      "ActorKind": {
        "enum": [
          "human",
          "codex",
          "deepseek",
          "claude",
          "local_model",
          "system"
        ],
        "title": "ActorKind",
        "type": "string"
      },
      "Approval": {
        "additionalProperties": false,
        "properties": {
          "actor_kind": {
            "const": "human",
            "title": "Actor Kind",
            "type": "string"
          },
          "approved_by": {
            "minLength": 1,
            "title": "Approved By",
            "type": "string"
          },
          "approved_content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Approved Content Hash",
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/FilmEntityRef"
          },
          "review_id": {
            "format": "uuid4",
            "title": "Review Id",
            "type": "string"
          },
          "target_id": {
            "format": "uuid4",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "target_id",
          "review_id",
          "actor_kind",
          "approved_by",
          "approved_content_hash"
        ],
        "title": "Approval",
        "type": "object"
      },
      "ApprovalCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_kind": {
            "$ref": "#/components/schemas/ActorKind"
          },
          "approved_by": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Approved By",
            "type": "string"
          },
          "candidate": {
            "$ref": "#/components/schemas/EntityVersionGuard"
          },
          "passed_review": {
            "$ref": "#/components/schemas/EntityVersionGuard"
          },
          "write": {
            "$ref": "#/components/schemas/CreateTargetGuard"
          }
        },
        "required": [
          "write",
          "actor_kind",
          "candidate",
          "passed_review",
          "approved_by"
        ],
        "title": "ApprovalCreateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Sidecar Film Core contract. Yingce remains authoritative for generic Host Project, ProjectUnit, Shot, Asset, Workflow and Task entities.",
    "title": "FilmOS Studio Film Core API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/approvals": {
      "post": {
        "operationId": "filmApprovalCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approval"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approval Create",
        "x-implementation-state": "implemented"
      }
    },
    "/audit-events": {
      "get": {
        "operationId": "filmAuditEventsGet",
        "parameters": [
          {
            "in": "query",
            "name": "targetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid4",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response Filmauditeventsget",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Events",
        "x-implementation-state": "implemented"
      }
    },
    "/commands/apply": {
      "post": {
        "operationId": "filmCommandApply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CreateEntityCommand"
                  },
                  {
                    "$ref": "#/components/schemas/SetStatesCommand"
                  }
                ],
                "title": "Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandApplyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command Apply",
        "x-implementation-state": "implemented"
      }
    },
    "/commands/preview": {
      "post": {
        "operationId": "filmCommandPreview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CreateEntityCommand"
                  },
                  {
                    "$ref": "#/components/schemas/SetStatesCommand"
                  }
                ],
                "title": "Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandPreviewResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command Preview",
        "x-implementation-state": "implemented"
      }
    },
    "/continuity/check": {
      "post": {
        "operationId": "filmContinuityCheck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContinuityCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContinuityCheckResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Continuity Check",
        "x-implementation-state": "implemented"
      }
    },
    "/entities/{filmEntityId}": {
      "get": {
        "operationId": "filmEntityGet",
        "parameters": [
          {
            "in": "path",
            "name": "filmEntityId",
            "required": true,
            "schema": {
              "format": "uuid4",
              "title": "Filmentityid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FilmProjectExtension"
                    },
                    {
                      "$ref": "#/components/schemas/ContentUnitExtension"
                    },
                    {
                      "$ref": "#/components/schemas/ShotExtension"
                    }
                  ],
                  "title": "Response Filmentityget"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entity Get",
        "x-implementation-state": "implemented"
      }
    },
    "/formal-records": {
      "post": {
        "operationId": "filmFormalRecordCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormalRecordCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormalRecordApplyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Formal Record Create",
        "x-implementation-state": "implemented"
      }
    },
    "/formal-records/{filmEntityId}": {
      "get": {
        "operationId": "filmFormalRecordGet",
        "parameters": [
          {
            "in": "path",
            "name": "filmEntityId",
            "required": true,
            "schema": {
              "format": "uuid4",
              "title": "Filmentityid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ScriptVersion"
                    },
                    {
                      "$ref": "#/components/schemas/ScriptDecision"
                    },
                    {
                      "$ref": "#/components/schemas/DirectorUnit"
                    },
                    {
                      "$ref": "#/components/schemas/CoverageLink"
                    },
                    {
                      "$ref": "#/components/schemas/VisualLockSet"
                    },
                    {
                      "$ref": "#/components/schemas/AssetBinding"
                    },
                    {
                      "$ref": "#/components/schemas/PromptDraft"
                    },
                    {
                      "$ref": "#/components/schemas/PromptDraftProvenance"
                    },
                    {
                      "$ref": "#/components/schemas/GenerationPackage"
                    },
                    {
                      "$ref": "#/components/schemas/GenerationAttemptEvidence"
                    },
                    {
                      "$ref": "#/components/schemas/Candidate"
                    },
                    {
                      "$ref": "#/components/schemas/Review"
                    },
                    {
                      "$ref": "#/components/schemas/Approval"
                    },
                    {
                      "$ref": "#/components/schemas/ContinuityCheckResult"
                    },
                    {
                      "$ref": "#/components/schemas/SceneTwinVersion"
                    },
                    {
                      "$ref": "#/components/schemas/CameraVersion"
                    },
                    {
                      "$ref": "#/components/schemas/BlockingVersion"
                    },
                    {
                      "$ref": "#/components/schemas/CompositionVersion"
                    }
                  ],
                  "title": "Response Filmformalrecordget"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Formal Record Get",
        "x-implementation-state": "implemented"
      }
    },
    "/generation-production/acceptance-authority": {
      "post": {
        "operationId": "filmGenerationProductionAcceptanceAuthorityPersist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Filmgenerationproductionacceptanceauthoritypersist",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generation Production Acceptance Authority Persist",
        "x-implementation-state": "implemented"
      }
    },
...
  },
  "servers": [
    {
      "url": "/film"
    }
  ]
}